schema update

+++ src/main/sql/schema.sql
@@ -23,7 +23,7 @@ create table comment (
 ) engine = InnoDb;
 
 create table article (
-    id bigint unsigned not null auto_increment primary key,
+    id int unsigned not null auto_increment primary key,
     name varchar(100) not null,
     category varchar(50),
     title varchar(250) not null,
@@ -31,15 +31,15 @@ create table article (
     deck varchar(250),
     description text(4000),
     keywords text(4000),
-    comment_target_id bigint unsigned not null unique,
+    comment_target_id int unsigned not null unique,
     date_created timestamp default 0,
     date_modified timestamp default current_timestamp on update current_timestamp,
     foreign key (comment_target_id) references comment_target (id)
 ) engine = InnoDb;
 
 create table article_page (
-    id bigint unsigned not null auto_increment primary key,
-    article_id bigint unsigned not null,
+    id int unsigned not null auto_increment primary key,
+    article_id int unsigned not null,
     page_number int(4) not null,
     title varchar(250) not null,
     date_created timestamp default 0,
